Moved Helper and Label to bin/util
