+++ lib/cli/index.js
@@ -1,7 +1,7 @@
 const table3 = require("cli-table3"),
-  chart = require("./chart"),
-  { labels, randomSay } = require("./label"),
-  { color, txt, table, formatNumber, ucfirst } = require("./helper"),
+  chart = require("../../bin/util/chart"),
+  { labels, randomSay } = require("../../bin/util/label"),
+  { color, txt, table, formatNumber, ucfirst } = require("../../bin/util/helper"),
   { line, br, space } = txt,
   { borders, col5 } = table,
   {
